requirements.txt
Django==5.2.5
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.10
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.1
